# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TechVerse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List portfolio projects
      parameters:
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: boolean
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List marketplace services
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{serviceId}:
    get:
      operationId: getService
      tags: [services]
      parameters:
        - in: path
          name: serviceId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    put:
      operationId: updateService
      tags: [services]
      parameters:
        - in: path
          name: serviceId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      parameters:
        - in: path
          name: serviceId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      parameters:
        - in: query
          name: serviceId
          schema:
            type: integer
        - in: query
          name: userId
          schema:
            type: string
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Admin
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        bio:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        github:
          type: string
          nullable: true
        linkedin:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required: [id, clerkId, name, email, skills, isAdmin, createdAt]
    UpdateUserProfileBody:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        avatar:
          type: string
        location:
          type: string
        website:
          type: string
        github:
          type: string
        linkedin:
          type: string
        skills:
          type: array
          items:
            type: string
...
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        userName:
          type: string
        userAvatar:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
          nullable: true
        techStack:
          type: array
          items:
            type: string
        category:
          type: string
        imageUrl:
          type: string
          nullable: true
        demoUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        featured:
          type: boolean
        views:
          type: integer
        likes:
          type: integer
        createdAt:
          type: string
      required: [id, userId, userName, title, description, techStack, category, featured, views, likes, createdAt]
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
        techStack:
          type: array
          items:
            type: string
        category:
          type: string
        imageUrl:
          type: string
        demoUrl:
          type: string
        githubUrl:
          type: string
        featured:
          type: boolean
      required: [title, description, techStack, category]
    Service:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        userName:
          type: string
        userAvatar:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        priceUnit:
          type: string
        deliveryDays:
          type: integer
        imageUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, userName, title, description, category, price, priceUnit, deliveryDays, tags, rating, reviewCount, isActive, createdAt]
    CreateServiceBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        priceUnit:
          type: string
        deliveryDays:
          type: integer
        imageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [title, description, category, price, priceUnit, deliveryDays]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        serviceId:
          type: integer
        bookingId:
          type: integer
        reviewerId:
          type: string
        reviewerName:
          type: string
        reviewerAvatar:
          type: string
          nullable: true
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
      required: [id, serviceId, bookingId, reviewerId, reviewerName, rating, comment, createdAt]
    CreateReviewBody:
      type: object
      properties:
        serviceId:
          type: integer
        bookingId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
      required: [serviceId, bookingId, rating, comment]
...